+++ src/main/java/com/tienda/entities/Usuario.java
@@ -22,7 +22,7 @@ public class Usuario implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idUsuario;    
+    private Integer idUsuario;    
     @NotEmpty
     private String username;    
     @NotEmpty
